requirements.txt
numpy==1.15.0
pandas==0.23.3
Pillow==5.2.0
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
